Store and enter credentials.
When https://is.stuba.sk/auth is opened, automatically enter stored
username and password. If user wants to store them for the first time,
clicking extension icon will open a popup with two text fields for
username and password in which user can enter them and click on Save
button to store them.